Attach policy actions to each result
